First draft of issue template for myDevelopment setup
Addresses #102 but will likely need a few revisions. I've combined the session setup and email scheduling into a single
issue here. Couldn't find a method of previewing these templates but have validated the YAML and can check the schema is
valid on the branch on GitHub before requesting a merge.